Move to FuzzyFinder since this is not limited to fzf anymore
